+++ src/transactions/manage_asset/UpdateAssetOpFrame.cpp
@@ -143,7 +143,7 @@ bool UpdateAssetOpFrame::doApply(Application& app, StorageHelper& storageHelper,
     }
     else
     {
-        if (app.getLedgerManager().shouldUse(LedgerVersion::FIX_AUTOREVIEW))
+        if (ledgerManager.shouldUse(LedgerVersion::FIX_AUTOREVIEW))
         {
             autoreview = false;
         }