+++ src/transactions/test/SaleWithSeveralQuotesTests.cpp
@@ -1,10 +1,8 @@
 #include <ledger/OfferHelper.h>
-#include <ledger/AssetPairHelper.h>
 #include <transactions/managers/FeeManager.h>
 #include <ledger/FeeHelper.h>
 #include <transactions/test/test_helper/ManageAccountRuleTestHelper.h>
 #include <transactions/test/test_helper/ManageAccountRoleTestHelper.h>
-#include "ledger/LedgerDeltaImpl.h"
 #include "ledger/ReviewableRequestHelper.h"
 #include "test/test.h"
 #include "TxTests.h"
@@ -14,16 +12,12 @@
 #include "test_helper/CreateAccountTestHelper.h"
 #include "test_helper/SaleRequestHelper.h"
 #include "test_helper/IssuanceRequestHelper.h"
-#include "test_helper/ManageBalanceTestHelper.h"
 #include "test_helper/ManageSaleTestHelper.h"
 #include "test_helper/ParticipateInSaleTestHelper.h"
-#include "transactions/dex/OfferManager.h"
 #include "ledger/SaleHelper.h"
 #include "test_helper/CheckSaleStateTestHelper.h"
 #include "test_helper/ReviewAssetRequestHelper.h"
 #include "test_helper/ReviewSaleRequestHelper.h"
-#include "test_helper/ReviewUpdateSaleDetailsRequestHelper.h"
-#include "test_helper/SetFeesTestHelper.h"
 #include "test/test_marshaler.h"
 #include "transactions/ManageKeyValueOpFrame.h"
 #include "test_helper/ManageKeyValueTestHelper.h"