+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 * multiple time reviewing of movement request
+* create and update asset's operations don't return empty details if update request sent
 
 ## 3.6.3